pyproject.toml
[project]
name = "spotter-task-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "openrouteservice>=2.3.3",
    "python-decouple>=3.8",
    "requests>=2.32.4",
]
